pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "xylo"
version = "0.1.0"
description = "App framework"
authors = ["Steve G <steve@pagekey.io>"]
packages = [{ include = "xylo", from="src" }]

[tool.poetry.scripts]
xylo = "xylo.cli:cli_entrypoint"

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.1.7"
cookiecutter = "^2.6.0"
pyyaml = "^6.0.2"
pydantic = "^2.9.1"

[tool.poetry.dev-dependencies]
ruff = "^0.6.5"
